moving maphelper to its own file as list of functions
